LinuxProtocolHandlerRegistrar: don't register absolute path
The absolute path would point at the unwrapped binary. Some distros such as
Nixpkgs must wrap their C# binaries in order to run them properly. This makes it
so that BSMM is taken from the $PATH which is up to the user to popuate with the
correct BeatSaberModManager.

Fixes https://github.com/affederaffe/BeatSaberModManager/issues/23